requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
blinker==1.4
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.1
docx2pdf==0.1.8
docxcompose==1.3.4
docxtpl==0.16.0
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.78.0
fastapi-mail==1.0.9
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
packaging==21.3
pdfkit==1.0.0
pydantic==1.9.1
pymongo==4.1.1
pyparsing==3.0.9
python-docx==0.8.11
python-dotenv==0.20.0
python-multipart==0.0.5
pywin32==304
redis==4.3.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.2.0
uvicorn==0.17.6
wrapt==1.14.1
